requirements.txt
biopython==1.76
numpy==1.18.4
pkg-resources==0.0.0
python-dotenv==0.13.0
